Crate socket_addr

A serialisable SocketAddr wrapper. There is no intent here to provide many more traits (perhaps serde). Deref is implemented to allow easy access to the underlying net::SocketAddr

Structs

SocketAddr
Wrapper around std::net::SocketAddr to enable it to be encoded and decoded.

SocketAddrV4
Utility struct of SocketAddrV4 for hole punching

SocketAddrV6
Utility struct of SocketAddrV6 for hole punching

Implementations